🏟🌄 Checkpoint ./src/presenters/pages/index.jsx:2/1952 ./src/presenters/includes/what-is-glitch.jsx:2/1795 ./src/presenters/includes/how-it-works.jsx:2/127
