+++ src/presenters/pages/index.jsx
@@ -9,54 +9,10 @@ import Link from '../includes/link.jsx';
 
 import Categories from '../categories.jsx';
 import Featured from '../featured.jsx';
-import OverlayVideo from '../overlays/overlay-video.jsx';
 import Questions from '../questions.jsx';
 import RandomCategories from '../random-categories.jsx';
 import RecentProjects from '../recent-projects.jsx';
-
-function loadScript(src) {
-  const script = document.createElement('script');
-  script.src = src;
-  script.async = true;
-  document.head.appendChild(script);
-}
-  
-class WhatIsGlitch extends React.Component {
-  componentDidMount() {
-    loadScript('//fast.wistia.com/embed/medias/i0m98yntdb.jsonp');
-    loadScript('//fast.wistia.com/assets/external/E-v1.js');
-  }
-  
-  render() {
-    const free = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Ffree.svg?1499350845981";
-    const play = "https://cdn.hyperdev.com/6ce807b5-7214-49d7-aadd-f11803bc35fd%2Fplay.svg";
-    const whatsGlitchWide = "https://cdn.glitch.com/f7224274-1330-4022-a8f2-8ae09dbd68a8%2Fwhats-glitch-wide.svg?1499885209761";
-    const whatsGlitchNarrow = "https://cdn.glitch.com/f7224274-1330-4022-a8f2-8ae09dbd68a8%2Fwhats-glitch-narrow.svg?1499884900667";
-    const whatsGlitchAlt = "Create a node app, or remix one. It updates as you type. Code with Friends!";
-    return (
-      <section className="what-is-glitch">
-        <h2>How It Works</h2>
-        <span>
-          <Link to="/about">
-            <figure title="How Glitch works">
-              <img className="wide" src={whatsGlitchWide} alt={whatsGlitchAlt}/>
-              <img className="narrow" src={whatsGlitchNarrow} alt={whatsGlitchAlt}/>
-            </figure>
-          </Link>
-          <div>
-            And it's <img className="free" src={free} alt="free"/>.{' '}
-            <OverlayVideo>
-              <div className="button video">
-                <img className="play-button" src={play} alt="play"/>
-                <span>How it works in 1 minute</span>
-              </div>
-            </OverlayVideo>
-          </div>
-        </span>
-      </section>
-    );
-  }
-}
+import HowItWorks from '../includes/how-it-works.jsx';
 
 const MadeInGlitch = () => (
   <section className="made-in-glitch">
@@ -78,7 +34,7 @@ const IndexPage = ({api, user}) => (
     <Featured/>
     <RandomCategories api={api}/>
     <Categories/>
-    {!(user && user.login) && <WhatIsGlitch/>}
+    {!(user && user.login) && <HowItWorks/>}
     <MadeInGlitch/>
   </main>
 );
